Fix problem with file paths.
In a log produced on TeX Live, file paths are split at 74 bytes instead of 79,
but the remaining lines are normally split at 79. This caused issues when
TW.fileExists() was defined and returned DOESNTEXIST -- the last part of the
file name (after the newline) was ignored and no path was captured.